Refactor reviewsummary route
Now routes to /api/reviews/{product_id}/summary
