+++ package.json
@@ -6,9 +6,10 @@
   "license": "",
   "scripts": {
     "build": "webpack",
+    "start": "node server/server.js",
+    "start:dev": "nodemon server/server.js",
     "watch": "webpack --mode=development --watch",
     "db:seed": "node database/seed/seedRandomData.js",
-    "server:dev": "nodemon server/server.js",
     "test:client": "jest -c=./tests/client/jest.client.config.js",
     "test:client:watch": "jest -c=./tests/client/jest.client.config.js --watch",
     "test:db": "jest -c=./tests/database/jest.db.config.js --ci --runInBand",
