+++ server/app.js
@@ -10,7 +10,6 @@ app.use((req, res, next) => {
 
 app.use('/', express.static(path.join(__dirname, '../public')));
 app.use('/api/reviews', require('./routes/reviews.js'));
-app.use('/api/reviewsummary', require('./routes/reviewsummary.js'));
 
 app.get('*', (req, res) => {
   res.sendFile(path.join(__dirname, '../public/index.html'));
