+++ tests/server/routes/reviews.mock.js
@@ -1,18 +1,9 @@
 const express = require('express');
-const reviewsTable = require('../../../database/methods/reviews.js');
+const { getReviews } = require('../../../database/methods/reviews.js');
+const { getReviewSummary } = require('../../../database/methods/reviewsummary.js');
+const { queryReviewRating } = require('../middleware/queryParams.mock.js');
 
 const router = express.Router();
-
-const queryReviewRating = (req, res, next) => {
-  if (req.query.review_rating !== undefined) {
-    if (Number.isNaN(Number(req.query.review_rating)) || req.query.review_rating === '') {
-      res.status(400).send('Bad Request.');
-      return;
-    }
-    req.options.review_rating = Number(req.query.review_rating);
-  }
-  next();
-};
 
 router.param('product_id', (req, res, next, product_id) => {
   if (Number.isNaN(Number(product_id))) {
@@ -23,6 +14,18 @@ router.param('product_id', (req, res, next, product_id) => {
   next();
 });
 
+router.route('/:product_id/summary')
+  .get(async (req, res) => {
+    try {
+      const reviewSummary = await getReviewSummary(req.options.product_id);
+      if (req.options.product_id === 9999) throw new Error('Mock Fail');
+      if (reviewSummary.length > 0) res.json(reviewSummary);
+      else res.status(404).send('Review Summary Not Found.');
+    } catch {
+      res.status(500).send('Internal Server Error.');
+    }
+  });
+
 router.route('/:product_id')
   .get(queryReviewRating, async (req, res) => {
     if (req.query.limit !== undefined) {
@@ -32,7 +35,7 @@ router.route('/:product_id')
       }
     } else req.query.limit = 0;
     try {
-      const reviews = await reviewsTable.getReviews(req.options, Number(req.query.limit));
+      const reviews = await getReviews(req.options, Number(req.query.limit));
       if (req.options.product_id === 9999) throw new Error('Mock Fail');
       if (reviews.length > 0) res.json(reviews);
       else res.status(404).send('Reviews Not Found.');
